Rename styled list
